shop/server/requirements.txt
fastapi==0.68.1
pydantic==1.10.2
uvicorn[standard]==0.15.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.5
bcrypt==3.2.0
sqlmodel==0.0.8
PyJWT==2.1.0
python-dotenv==1.0.0
